MS-341-344: atanh/asinh/expm1/log1p/log10/erfc bwd parity + bench 152 + type stubs
## Parity tests: PyTorch 286 (+6), JAX 142 (+3)
PyTorch:
- test_atanh_bwd: d/dx atanh(x) = 1/(1-x^2)
- test_asinh_bwd: d/dx asinh(x) = 1/sqrt(x^2+1)
- test_expm1_bwd: d/dx expm1(x) = exp(x)
- test_log1p_bwd: d/dx log1p(x) = 1/(1+x)
- test_log10_bwd: d/dx log10(x) = 1/(x*ln10)
- test_erfc_bwd: d/dx erfc(x) = -(2/sqrt(pi))*exp(-x^2)

JAX: test_atanh_bwd / test_expm1_bwd / test_log1p_bwd

## Type stubs (MS-343)
pycoeus.pyi: Added stubs for 16 new ops:
tan, asin, acos, atan, sinh, cosh, tanh_direct, atanh, asinh, acosh,
log2, log10, log1p, exp2, expm1, selu

## G-043 bench: 144 -> 152 rows
atanh_bwd, asinh_bwd, acosh_bwd, expm1_bwd, log1p_bwd,
log10_bwd, acos_bwd, asin_bwd (all fwd+bwd benchmarks)

65/65 coeus-autograd + 190/190 coeus-ops tests passing.
