+++ coeus-nn/src/activation/mod.rs
@@ -1,0 +1,5 @@
+pub mod basic;
+pub mod parametric;
+
+pub use basic::*;
+pub use parametric::*;
