+++ coeus-python/pycoeus.pyi
@@ -221,6 +221,22 @@ def trunc(input: Tensor) -> Tensor: ...
 def sin(input: Tensor) -> Tensor: ...
 def cos(input: Tensor) -> Tensor: ...
 def pow(input: Tensor, exp: float) -> Tensor: ...
+def tan(input: Tensor) -> Tensor: ...
+def asin(input: Tensor) -> Tensor: ...
+def acos(input: Tensor) -> Tensor: ...
+def atan(input: Tensor) -> Tensor: ...
+def sinh(input: Tensor) -> Tensor: ...
+def cosh(input: Tensor) -> Tensor: ...
+def tanh_direct(input: Tensor) -> Tensor: ...
+def atanh(input: Tensor) -> Tensor: ...
+def asinh(input: Tensor) -> Tensor: ...
+def acosh(input: Tensor) -> Tensor: ...
+def log2(input: Tensor) -> Tensor: ...
+def log10(input: Tensor) -> Tensor: ...
+def log1p(input: Tensor) -> Tensor: ...
+def exp2(input: Tensor) -> Tensor: ...
+def expm1(input: Tensor) -> Tensor: ...
+def selu(input: Tensor) -> Tensor: ...
 def clamp(input: Tensor, min_val: float, max_val: float) -> Tensor: ...
 def fft(input: Tensor) -> ComplexTensor: ...
 def ifft(input: ComplexTensor) -> Tensor: ...
